Crate rusoto_cloudformation

AWS CloudFormation
AWS CloudFormation allows you to create and manage AWS infrastructure deployments predictably and repeatedly. You can use AWS CloudFormation to leverage AWS products, such as Amazon Elastic Compute Cloud, Amazon Elastic Block Store, Amazon Simple Notification Service, Elastic Load Balancing, and Auto Scaling to build highly-reliable, highly scalable, cost-effective applications without creating or configuring the underlying AWS infrastructure.

With AWS CloudFormation, you declare all of your resources and dependencies in a template file. The template defines a collection of resources as a single unit called a stack. AWS CloudFormation creates and deletes all member resources of the stack together and manages all dependencies between the resources for you.

For more information about AWS CloudFormation, see the AWS CloudFormation Product Page.

Amazon CloudFormation makes use of other AWS products. If you need additional technical information about a specific AWS product, you can find the product's technical documentation at docs.aws.amazon.com.

If you're using the service, you're probably looking for CloudFormationClient and CloudFormation.

Structs

AccountGateResult
Structure that contains the results of the account gate function which AWS CloudFormation invokes, if present, before proceeding with a stack set operation in an account and region.

For each account and region, AWS CloudFormation lets you specify a Lamdba function that encapsulates any requirements that must be met before CloudFormation can proceed with a stack set operation in that account and region. CloudFormation invokes the function each time a stack set operation is requested for that account and region; if the function returns FAILED, CloudFormation cancels the operation in that account and region, and sets the stack set operation result status for that account and region to FAILED.

For more information, see Configuring a target account gate.

AccountLimit
The AccountLimit data type. For more information about account limits, see AWS CloudFormation Limits in the AWS CloudFormation User Guide.

CancelUpdateStackInput
The input for the CancelUpdateStack action.

Change
The Change structure describes the changes AWS CloudFormation will perform if you execute the change set.

ChangeSetSummary
The ChangeSetSummary structure describes a change set, its status, and the stack with which it's associated.

CloudFormationClient
A client for the AWS CloudFormation API.

ContinueUpdateRollbackInput
The input for the ContinueUpdateRollback action.

ContinueUpdateRollbackOutput
The output for a ContinueUpdateRollback action.

CreateChangeSetInput
The input for the CreateChangeSet action.

CreateChangeSetOutput
The output for the CreateChangeSet action.

CreateStackInput
The input for CreateStack action.

CreateStackInstancesInput

CreateStackInstancesOutput

CreateStackOutput
The output for a CreateStack action.

CreateStackSetInput

CreateStackSetOutput

DeleteChangeSetInput
The input for the DeleteChangeSet action.

DeleteChangeSetOutput
The output for the DeleteChangeSet action.

DeleteStackInput
The input for DeleteStack action.

DeleteStackInstancesInput

DeleteStackInstancesOutput

DeleteStackSetInput

DeleteStackSetOutput

DescribeAccountLimitsInput
The input for the DescribeAccountLimits action.

DescribeAccountLimitsOutput
The output for the DescribeAccountLimits action.

DescribeChangeSetInput
The input for the DescribeChangeSet action.

DescribeChangeSetOutput
The output for the DescribeChangeSet action.

DescribeStackDriftDetectionStatusInput

DescribeStackDriftDetectionStatusOutput

DescribeStackEventsInput
The input for DescribeStackEvents action.

DescribeStackEventsOutput
The output for a DescribeStackEvents action.

DescribeStackInstanceInput

DescribeStackInstanceOutput

DescribeStackResourceDriftsInput

DescribeStackResourceDriftsOutput

DescribeStackResourceInput
The input for DescribeStackResource action.

DescribeStackResourceOutput
The output for a DescribeStackResource action.

DescribeStackResourcesInput
The input for DescribeStackResources action.

DescribeStackResourcesOutput
The output for a DescribeStackResources action.

DescribeStackSetInput

DescribeStackSetOperationInput

DescribeStackSetOperationOutput

DescribeStackSetOutput

DescribeStacksInput
The input for DescribeStacks action.

DescribeStacksOutput
The output for a DescribeStacks action.

DetectStackDriftInput

DetectStackDriftOutput

DetectStackResourceDriftInput

DetectStackResourceDriftOutput

EstimateTemplateCostInput
The input for an EstimateTemplateCost action.

EstimateTemplateCostOutput
The output for a EstimateTemplateCost action.

ExecuteChangeSetInput
The input for the ExecuteChangeSet action.

ExecuteChangeSetOutput
The output for the ExecuteChangeSet action.

Export
The Export structure describes the exported output values for a stack.

GetStackPolicyInput
The input for the GetStackPolicy action.

GetStackPolicyOutput
The output for the GetStackPolicy action.

GetTemplateInput
The input for a GetTemplate action.

GetTemplateOutput
The output for GetTemplate action.

GetTemplateSummaryInput
The input for the GetTemplateSummary action.

GetTemplateSummaryOutput
The output for the GetTemplateSummary action.

ListChangeSetsInput
The input for the ListChangeSets action.

ListChangeSetsOutput
The output for the ListChangeSets action.

ListExportsInput

ListExportsOutput

ListImportsInput

ListImportsOutput

ListStackInstancesInput

ListStackInstancesOutput

ListStackResourcesInput
The input for the ListStackResource action.

ListStackResourcesOutput
The output for a ListStackResources action.

ListStackSetOperationResultsInput

ListStackSetOperationResultsOutput

ListStackSetOperationsInput

ListStackSetOperationsOutput

ListStackSetsInput

ListStackSetsOutput

ListStacksInput
The input for ListStacks action.

ListStacksOutput
The output for ListStacks action.

Output
The Output data type.

Parameter
The Parameter data type.

ParameterConstraints
A set of criteria that AWS CloudFormation uses to validate parameter values. Although other constraints might be defined in the stack template, AWS CloudFormation returns only the AllowedValues property.

ParameterDeclaration
The ParameterDeclaration data type.

PhysicalResourceIdContextKeyValuePair
Context information that enables AWS CloudFormation to uniquely identify a resource. AWS CloudFormation uses context key-value pairs in cases where a resource's logical and physical IDs are not enough to uniquely identify that resource. Each context key-value pair specifies a resource that contains the targeted resource.

PropertyDifference
Information about a resource property whose actual value differs from its expected value, as defined in the stack template and any values specified as template parameters. These will be present only for resources whose StackResourceDriftStatus is MODIFIED. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

ResourceChange
The ResourceChange structure describes the resource and the action that AWS CloudFormation will perform on it if you execute this change set.

ResourceChangeDetail
For a resource with Modify as the action, the ResourceChange structure describes the changes AWS CloudFormation will make to that resource.

ResourceTargetDefinition
The field that AWS CloudFormation will change, such as the name of a resource's property, and whether the resource will be recreated.

RollbackConfiguration
Structure containing the rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

Rollback triggers enable you to have AWS CloudFormation monitor the state of your application during stack creation and updating, and to roll back that operation if the application breaches the threshold of any of the alarms you've specified. For more information, see Monitor and Roll Back Stack Operations.

RollbackTrigger
A rollback trigger AWS CloudFormation monitors during creation and updating of stacks. If any of the alarms you specify goes to ALARM state during the stack operation or within the specified monitoring period afterwards, CloudFormation rolls back the entire stack operation.

SetStackPolicyInput
The input for the SetStackPolicy action.

SignalResourceInput
The input for the SignalResource action.

Stack
The Stack data type.

StackDriftInformation
Contains information about whether the stack's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. A stack is considered to have drifted if one or more of its resources have drifted.

StackDriftInformationSummary
Contains information about whether the stack's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. A stack is considered to have drifted if one or more of its resources have drifted.

StackEvent
The StackEvent data type.

StackInstance
An AWS CloudFormation stack, in a specific account and region, that's part of a stack set operation. A stack instance is a reference to an attempted or actual stack in a given account within a given region. A stack instance can exist without a stack—for example, if the stack couldn't be created for some reason. A stack instance is associated with only one stack set. Each stack instance contains the ID of its associated stack set, as well as the ID of the actual stack and the stack status.

StackInstanceSummary
The structure that contains summary information about a stack instance.

StackResource
The StackResource data type.

StackResourceDetail
Contains detailed information about the specified stack resource.

StackResourceDrift
Contains the drift information for a resource that has been checked for drift. This includes actual and expected property values for resources in which AWS CloudFormation has detected drift. Only resource properties explicitly defined in the stack template are checked for drift. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

Resources that do not currently support drift detection cannot be checked. For a list of resources that support drift detection, see Resources that Support Drift Detection.

Use DetectStackResourceDrift to detect drift on individual resources, or DetectStackDrift to detect drift on all resources in a given stack that support drift detection.

StackResourceDriftInformation
Contains information about whether the resource's actual configuration differs, or has drifted, from its expected configuration.

StackResourceDriftInformationSummary
Summarizes information about whether the resource's actual configuration differs, or has drifted, from its expected configuration.

StackResourceSummary
Contains high-level information about the specified stack resource.

StackSet
A structure that contains information about a stack set. A stack set enables you to provision stacks into AWS accounts and across regions by using a single CloudFormation template. In the stack set, you specify the template to use, as well as any parameters and capabilities that the template requires.

StackSetOperation
The structure that contains information about a stack set operation.

StackSetOperationPreferences
The user-specified preferences for how AWS CloudFormation performs a stack set operation.

For more information on maximum concurrent accounts and failure tolerance, see Stack set operation options.

StackSetOperationResultSummary
The structure that contains information about a specified operation's results for a given account in a given region.

StackSetOperationSummary
The structures that contain summary information about the specified operation.

StackSetSummary
The structures that contain summary information about the specified stack set.

StackSummary
The StackSummary Data Type

StopStackSetOperationInput

StopStackSetOperationOutput

Tag
The Tag type enables you to specify a key-value pair that can be used to store information about an AWS CloudFormation stack.

TemplateParameter
The TemplateParameter data type.

UpdateStackInput
The input for an UpdateStack action.

UpdateStackInstancesInput

UpdateStackInstancesOutput

UpdateStackOutput
The output for an UpdateStack action.

UpdateStackSetInput

UpdateStackSetOutput

UpdateTerminationProtectionInput

UpdateTerminationProtectionOutput

ValidateTemplateInput
The input for ValidateTemplate action.

ValidateTemplateOutput
The output for ValidateTemplate action.

Enums

CancelUpdateStackError
Errors returned by CancelUpdateStack

ContinueUpdateRollbackError
Errors returned by ContinueUpdateRollback

CreateChangeSetError
Errors returned by CreateChangeSet

CreateStackError
Errors returned by CreateStack

CreateStackInstancesError
Errors returned by CreateStackInstances

CreateStackSetError
Errors returned by CreateStackSet

DeleteChangeSetError
Errors returned by DeleteChangeSet

DeleteStackError
Errors returned by DeleteStack

DeleteStackInstancesError
Errors returned by DeleteStackInstances

DeleteStackSetError
Errors returned by DeleteStackSet

DescribeAccountLimitsError
Errors returned by DescribeAccountLimits

DescribeChangeSetError
Errors returned by DescribeChangeSet

DescribeStackDriftDetectionStatusError
Errors returned by DescribeStackDriftDetectionStatus

DescribeStackEventsError
Errors returned by DescribeStackEvents

DescribeStackInstanceError
Errors returned by DescribeStackInstance

DescribeStackResourceDriftsError
Errors returned by DescribeStackResourceDrifts

DescribeStackResourceError
Errors returned by DescribeStackResource

DescribeStackResourcesError
Errors returned by DescribeStackResources

DescribeStackSetError
Errors returned by DescribeStackSet

DescribeStackSetOperationError
Errors returned by DescribeStackSetOperation

DescribeStacksError
Errors returned by DescribeStacks

DetectStackDriftError
Errors returned by DetectStackDrift

DetectStackResourceDriftError
Errors returned by DetectStackResourceDrift

EstimateTemplateCostError
Errors returned by EstimateTemplateCost

ExecuteChangeSetError
Errors returned by ExecuteChangeSet

GetStackPolicyError
Errors returned by GetStackPolicy

GetTemplateError
Errors returned by GetTemplate

GetTemplateSummaryError
Errors returned by GetTemplateSummary

ListChangeSetsError
Errors returned by ListChangeSets

ListExportsError
Errors returned by ListExports

ListImportsError
Errors returned by ListImports

ListStackInstancesError
Errors returned by ListStackInstances

ListStackResourcesError
Errors returned by ListStackResources

ListStackSetOperationResultsError
Errors returned by ListStackSetOperationResults

ListStackSetOperationsError
Errors returned by ListStackSetOperations

ListStackSetsError
Errors returned by ListStackSets

ListStacksError
Errors returned by ListStacks

SetStackPolicyError
Errors returned by SetStackPolicy

SignalResourceError
Errors returned by SignalResource

StopStackSetOperationError
Errors returned by StopStackSetOperation

UpdateStackError
Errors returned by UpdateStack

UpdateStackInstancesError
Errors returned by UpdateStackInstances

UpdateStackSetError
Errors returned by UpdateStackSet

UpdateTerminationProtectionError
Errors returned by UpdateTerminationProtection

ValidateTemplateError
Errors returned by ValidateTemplate

Traits

CloudFormation
Trait representing the capabilities of the AWS CloudFormation API. AWS CloudFormation clients implement this trait.