Struct rusoto_cloudformation::AccountGateResult

pub struct AccountGateResult {
    pub status: Option<String>,
    pub status_reason: Option<String>,
}

Structure that contains the results of the account gate function which AWS CloudFormation invokes, if present, before proceeding with a stack set operation in an account and region.

For each account and region, AWS CloudFormation lets you specify a Lamdba function that encapsulates any requirements that must be met before CloudFormation can proceed with a stack set operation in that account and region. CloudFormation invokes the function each time a stack set operation is requested for that account and region; if the function returns FAILED, CloudFormation cancels the operation in that account and region, and sets the stack set operation result status for that account and region to FAILED.

For more information, see Configuring a target account gate.

Fields
status: Option<String>

The status of the account gate function.

SUCCEEDED: The account gate function has determined that the account and region passes any requirements for a stack set operation to occur. AWS CloudFormation proceeds with the stack operation in that account and region.

FAILED: The account gate function has determined that the account and region does not meet the requirements for a stack set operation to occur. AWS CloudFormation cancels the stack set operation in that account and region, and sets the stack set operation result status for that account and region to FAILED.

SKIPPED: AWS CloudFormation has skipped calling the account gate function for this account and region, for one of the following reasons:

An account gate function has not been specified for the account and region. AWS CloudFormation proceeds with the stack set operation in this account and region.

The AWSCloudFormationStackSetExecutionRole of the stack set adminstration account lacks permissions to invoke the function. AWS CloudFormation proceeds with the stack set operation in this account and region.

Either no action is necessary, or no action is possible, on the stack. AWS CloudFormation skips the stack set operation in this account and region.

status_reason: Option<String>

The reason for the account gate status assigned to this account and region for the stack set operation.