Struct rusoto_cloudformation::CloudFormationClient

pub struct CloudFormationClient { /* fields omitted */ }

A client for the AWS CloudFormation API.

Methods

impl CloudFormationClient

pub fn new(region: Region) -> CloudFormationClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudFormationClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,