Struct rusoto_cloudformation::DeleteStackInput

pub struct DeleteStackInput {
    pub client_request_token: Option<String>,
    pub retain_resources: Option<Vec<String>>,
    pub role_arn: Option<String>,
    pub stack_name: String,
}

The input for DeleteStack action.

Fields
client_request_token: Option<String>

A unique identifier for this DeleteStack request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to delete a stack with the same name. You might retry DeleteStack requests to ensure that AWS CloudFormation successfully received them.

retain_resources: Option<Vec<String>>

For stacks in the DELETE_FAILED state, a list of resource logical IDs that are associated with the resources you want to retain. During deletion, AWS CloudFormation deletes the stack but does not delete the retained resources.

Retaining resources is useful when you cannot delete a resource, such as a non-empty S3 bucket, but you want to delete the stack.

role_arn: Option<String>

The Amazon Resource Name (ARN) of an AWS Identity and Access Management (IAM) role that AWS CloudFormation assumes to delete the stack. AWS CloudFormation uses the role's credentials to make calls on your behalf.

If you don't specify a value, AWS CloudFormation uses the role that was previously associated with the stack. If no role is available, AWS CloudFormation uses a temporary session that is generated from your user credentials.

stack_name: String

The name or the unique stack ID that is associated with the stack.