Struct rusoto_cloudformation::ResourceChange

pub struct ResourceChange {
    pub action: Option<String>,
    pub details: Option<Vec<ResourceChangeDetail>>,
    pub logical_resource_id: Option<String>,
    pub physical_resource_id: Option<String>,
    pub replacement: Option<String>,
    pub resource_type: Option<String>,
    pub scope: Option<Vec<String>>,
}

The ResourceChange structure describes the resource and the action that AWS CloudFormation will perform on it if you execute this change set.

Fields
action: Option<String>

The action that AWS CloudFormation takes on the resource, such as Add (adds a new resource), Modify (changes a resource), or Remove (deletes a resource).

details: Option<Vec<ResourceChangeDetail>>

For the Modify action, a list of ResourceChangeDetail structures that describes the changes that AWS CloudFormation will make to the resource.

logical_resource_id: Option<String>

The resource's logical ID, which is defined in the stack's template.

physical_resource_id: Option<String>

The resource's physical ID (resource name). Resources that you are adding don't have physical IDs because they haven't been created.

replacement: Option<String>

For the Modify action, indicates whether AWS CloudFormation will replace the resource by creating a new one and deleting the old one. This value depends on the value of the RequiresRecreation property in the ResourceTargetDefinition structure. For example, if the RequiresRecreation field is Always and the Evaluation field is Static, Replacement is True. If the RequiresRecreation field is Always and the Evaluation field is Dynamic, Replacement is Conditionally.

If you have multiple changes with different RequiresRecreation values, the Replacement value depends on the change with the most impact. A RequiresRecreation value of Always has the most impact, followed by Conditionally, and then Never.

resource_type: Option<String>

The type of AWS CloudFormation resource, such as AWS::S3::Bucket.

scope: Option<Vec<String>>

For the Modify action, indicates which resource attribute is triggering this update, such as a change in the resource attribute's Metadata, Properties, or Tags.