Struct rusoto_cloudformation::StackInstance

pub struct StackInstance {
    pub account: Option<String>,
    pub region: Option<String>,
    pub stack_id: Option<String>,
    pub stack_set_id: Option<String>,
    pub status: Option<String>,
    pub status_reason: Option<String>,
}

An AWS CloudFormation stack, in a specific account and region, that's part of a stack set operation. A stack instance is a reference to an attempted or actual stack in a given account within a given region. A stack instance can exist without a stack—for example, if the stack couldn't be created for some reason. A stack instance is associated with only one stack set. Each stack instance contains the ID of its associated stack set, as well as the ID of the actual stack and the stack status.

Fields
account: Option<String>

The name of the AWS account that the stack instance is associated with.

region: Option<String>

The name of the AWS region that the stack instance is associated with.

stack_id: Option<String>

The ID of the stack instance.

stack_set_id: Option<String>

The name or unique ID of the stack set that the stack instance is associated with.

status: Option<String>

The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because:

The associated stack failed during a CreateStackSet or UpdateStackSet operation.

The stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

status_reason: Option<String>

The explanation for the specific status code that is assigned to this stack instance.