Crate rusoto_cloudformation

AWS CloudFormation
AWS CloudFormation allows you to create and manage AWS infrastructure deployments predictably and repeatedly. You can use AWS CloudFormation to leverage AWS products, such as Amazon Elastic Compute Cloud, Amazon Elastic Block Store, Amazon Simple Notification Service, Elastic Load Balancing, and Auto Scaling to build highly-reliable, highly scalable, cost-effective applications without creating or configuring the underlying AWS infrastructure.

With AWS CloudFormation, you declare all of your resources and dependencies in a template file. The template defines a collection of resources as a single unit called a stack. AWS CloudFormation creates and deletes all member resources of the stack together and manages all dependencies between the resources for you.

For more information about AWS CloudFormation, see the AWS CloudFormation Product Page.

Amazon CloudFormation makes use of other AWS products. If you need additional technical information about a specific AWS product, you can find the product's technical documentation at docs.aws.amazon.com.

If you're using the service, you're probably looking for CloudFormationClient and CloudFormation.

Structs

AccountLimit
The AccountLimit data type.

CancelUpdateStackInput
The input for the CancelUpdateStack action.

Change
The Change structure describes the changes AWS CloudFormation will perform if you execute the change set.

ChangeSetSummary
The ChangeSetSummary structure describes a change set, its status, and the stack with which it's associated.

CloudFormationClient
A client for the AWS CloudFormation API.

ContinueUpdateRollbackInput
The input for the ContinueUpdateRollback action.

ContinueUpdateRollbackOutput
The output for a ContinueUpdateRollback action.

CreateChangeSetInput
The input for the CreateChangeSet action.

CreateChangeSetOutput
The output for the CreateChangeSet action.

CreateStackInput
The input for CreateStack action.

CreateStackOutput
The output for a CreateStack action.

DeleteChangeSetInput
The input for the DeleteChangeSet action.

DeleteChangeSetOutput
The output for the DeleteChangeSet action.

DeleteStackInput
The input for DeleteStack action.

DescribeAccountLimitsInput
The input for the DescribeAccountLimits action.

DescribeAccountLimitsOutput
The output for the DescribeAccountLimits action.

DescribeChangeSetInput
The input for the DescribeChangeSet action.

DescribeChangeSetOutput
The output for the DescribeChangeSet action.

DescribeStackEventsInput
The input for DescribeStackEvents action.

DescribeStackEventsOutput
The output for a DescribeStackEvents action.

DescribeStackResourceInput
The input for DescribeStackResource action.

DescribeStackResourceOutput
The output for a DescribeStackResource action.

DescribeStackResourcesInput
The input for DescribeStackResources action.

DescribeStackResourcesOutput
The output for a DescribeStackResources action.

DescribeStacksInput
The input for DescribeStacks action.

DescribeStacksOutput
The output for a DescribeStacks action.

EstimateTemplateCostInput
The input for an EstimateTemplateCost action.

EstimateTemplateCostOutput
The output for a EstimateTemplateCost action.

ExecuteChangeSetInput
The input for the ExecuteChangeSet action.

ExecuteChangeSetOutput
The output for the ExecuteChangeSet action.

Export
The Export structure describes the exported output values for a stack.

GetStackPolicyInput
The input for the GetStackPolicy action.

GetStackPolicyOutput
The output for the GetStackPolicy action.

GetTemplateInput
The input for a GetTemplate action.

GetTemplateOutput
The output for GetTemplate action.

GetTemplateSummaryInput
The input for the GetTemplateSummary action.

GetTemplateSummaryOutput
The output for the GetTemplateSummary action.

ListChangeSetsInput
The input for the ListChangeSets action.

ListChangeSetsOutput
The output for the ListChangeSets action.

ListExportsInput

ListExportsOutput

ListImportsInput

ListImportsOutput

ListStackResourcesInput
The input for the ListStackResource action.

ListStackResourcesOutput
The output for a ListStackResources action.

ListStacksInput
The input for ListStacks action.

ListStacksOutput
The output for ListStacks action.

Output
The Output data type.

Parameter
The Parameter data type.

ParameterConstraints
A set of criteria that AWS CloudFormation uses to validate parameter values. Although other constraints might be defined in the stack template, AWS CloudFormation returns only the AllowedValues property.

ParameterDeclaration
The ParameterDeclaration data type.

ResourceChange
The ResourceChange structure describes the resource and the action that AWS CloudFormation will perform on it if you execute this change set.

ResourceChangeDetail
For a resource with Modify as the action, the ResourceChange structure describes the changes AWS CloudFormation will make to that resource.

ResourceTargetDefinition
The field that AWS CloudFormation will change, such as the name of a resource's property, and whether the resource will be recreated.

SetStackPolicyInput
The input for the SetStackPolicy action.

SignalResourceInput
The input for the SignalResource action.

Stack
The Stack data type.

StackEvent
The StackEvent data type.

StackResource
The StackResource data type.

StackResourceDetail
Contains detailed information about the specified stack resource.

StackResourceSummary
Contains high-level information about the specified stack resource.

StackSummary
The StackSummary Data Type

Tag
The Tag type enables you to specify a key-value pair that can be used to store information about an AWS CloudFormation stack.

TemplateParameter
The TemplateParameter data type.

UpdateStackInput
The input for an UpdateStack action.

UpdateStackOutput
The output for an UpdateStack action.

ValidateTemplateInput
The input for ValidateTemplate action.

ValidateTemplateOutput
The output for ValidateTemplate action.

Enums

CancelUpdateStackError
Errors returned by CancelUpdateStack

ContinueUpdateRollbackError
Errors returned by ContinueUpdateRollback

CreateChangeSetError
Errors returned by CreateChangeSet

CreateStackError
Errors returned by CreateStack

DeleteChangeSetError
Errors returned by DeleteChangeSet

DeleteStackError
Errors returned by DeleteStack

DescribeAccountLimitsError
Errors returned by DescribeAccountLimits

DescribeChangeSetError
Errors returned by DescribeChangeSet

DescribeStackEventsError
Errors returned by DescribeStackEvents

DescribeStackResourceError
Errors returned by DescribeStackResource

DescribeStackResourcesError
Errors returned by DescribeStackResources

DescribeStacksError
Errors returned by DescribeStacks

EstimateTemplateCostError
Errors returned by EstimateTemplateCost

ExecuteChangeSetError
Errors returned by ExecuteChangeSet

GetStackPolicyError
Errors returned by GetStackPolicy

GetTemplateError
Errors returned by GetTemplate

GetTemplateSummaryError
Errors returned by GetTemplateSummary

ListChangeSetsError
Errors returned by ListChangeSets

ListExportsError
Errors returned by ListExports

ListImportsError
Errors returned by ListImports

ListStackResourcesError
Errors returned by ListStackResources

ListStacksError
Errors returned by ListStacks

SetStackPolicyError
Errors returned by SetStackPolicy

SignalResourceError
Errors returned by SignalResource

UpdateStackError
Errors returned by UpdateStack

ValidateTemplateError
Errors returned by ValidateTemplate

Traits

CloudFormation
Trait representing the capabilities of the AWS CloudFormation API. AWS CloudFormation clients implement this trait.