Struct rusoto_cloudformation::DescribeChangeSetOutput

pub struct DescribeChangeSetOutput {
    pub capabilities: Option<Vec<String>>,
    pub change_set_id: Option<String>,
    pub change_set_name: Option<String>,
    pub changes: Option<Vec<Change>>,
    pub creation_time: Option<String>,
    pub description: Option<String>,
    pub execution_status: Option<String>,
    pub next_token: Option<String>,
    pub notification_ar_ns: Option<Vec<String>>,
    pub parameters: Option<Vec<Parameter>>,
    pub rollback_configuration: Option<RollbackConfiguration>,
    pub stack_id: Option<String>,
    pub stack_name: Option<String>,
    pub status: Option<String>,
    pub status_reason: Option<String>,
    pub tags: Option<Vec<Tag>>,
}

The output for the DescribeChangeSet action.

Fields
capabilities: Option<Vec<String>>

If you execute the change set, the list of capabilities that were explicitly acknowledged when the change set was created.

change_set_id: Option<String>

The ARN of the change set.

change_set_name: Option<String>

The name of the change set.

changes: Option<Vec<Change>>

A list of Change structures that describes the resources AWS CloudFormation changes if you execute the change set.

creation_time: Option<String>

The start time when the change set was created, in UTC.

description: Option<String>

Information about the change set.

execution_status: Option<String>

If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because AWS CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

next_token: Option<String>

If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

notification_ar_ns: Option<Vec<String>>

The ARNs of the Amazon Simple Notification Service (Amazon SNS) topics that will be associated with the stack if you execute the change set.

parameters: Option<Vec<Parameter>>

A list of Parameter structures that describes the input parameters and their values used to create the change set. For more information, see the Parameter data type.

rollback_configuration: Option<RollbackConfiguration>

The rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

stack_id: Option<String>

The ARN of the stack that is associated with the change set.

stack_name: Option<String>

The name of the stack that is associated with the change set.

status: Option<String>

The current status of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

status_reason: Option<String>

A description of the change set's status. For example, if your attempt to create a change set failed, AWS CloudFormation shows the error message.

tags: Option<Vec<Tag>>

If you execute the change set, the tags that will be associated with the stack.